refactored few api and cli in compliance
